classification-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sentence-transformers==3.0.1
psycopg2-binary==2.9.9
pgvector==0.3.2
scikit-learn==1.5.1
hdbscan>=0.8.38
numpy==1.26.4
pydantic==2.8.0
apscheduler==3.10.4
pdfplumber==0.11.0
requests==2.32.3
